Fix second round of black-box agent findings
- verbose: -v/-vv now emit stderr diagnostics (HTTP method+URL, token
  refresh, SMTP host:port/mode, send + mark-seen events). New diag module;
  secrets are never logged. stdout stays pure JSON.
- seen: reading a message via `get` now PATCHes seen=true upstream
  (merge-patch, best-effort), so `read --unread` actually excludes
  already-read mail. `wait`/`otp` inherit this since they fetch via `get`.
- rm: Receiver::delete now returns whether the inbox still existed; the
  --handle path reports existed truthfully (404 -> false) instead of
  "delete call didn't error". Semantics documented on run_rm.
- read --limit: rejects 0/negative as a CONFIG error at parse time
  instead of silently returning []. Field widened to u64 with a 1.. range.

stdout/stderr duplication (finding #5) is by design: stdout is the machine
JSON envelope, stderr the human line — left unchanged.

49 unit + 10 integration tests pass; clippy/fmt clean.

+++ src/cli.rs
@@ -88,9 +88,9 @@ pub struct ReadArgs {
     /// Only show messages at/after this ISO-8601 timestamp.
     #[arg(long)]
     pub since: Option<String>,
-    /// Maximum number of messages to return.
-    #[arg(long, default_value_t = 50)]
-    pub limit: usize,
+    /// Maximum number of messages to return (must be >= 1).
+    #[arg(long, default_value_t = 50, value_parser = clap::value_parser!(u64).range(1..))]
+    pub limit: u64,
 }
 
 /// `tmail get <id|address> <msgId>` — or just `<msgId>` with `--handle`/one inbox.
